# refactor: improve list command performance