--- a/lib/aptible/cli/subcommands/apps.rb
+++ b/lib/aptible/cli/subcommands/apps.rb
@@ -17,7 +17,7 @@ module Aptible
                   'handle'
                 ) do |node|
                   acc_map = {}
-                  accounts = scoped_environments(options, Renderer.format!='json')
+                  accounts = scoped_environments(options)
                   accounts.each do |account|
                     acc_map[account.links.self.href] = account
                   end