--- a/lib/aptible/cli/subcommands/db.rb
+++ b/lib/aptible/cli/subcommands/db.rb
@@ -22,7 +22,7 @@ module Aptible
                   'handle'
                 ) do |node|
                   acc_map = {}
-                  accounts = scoped_environments(options, Renderer.format!='json')
+                  accounts = scoped_environments(options)
                   accounts.each do |account|
                     acc_map[account.links.self.href] = account
                   end