--- a/lib/aptible/cli/subcommands/metric_drain.rb
+++ b/lib/aptible/cli/subcommands/metric_drain.rb
@@ -25,11 +25,8 @@ module Aptible
                   { 'environment' => 'handle' },
                   'handle'
                 ) do |node|
-                  acc_map = {}
                   accounts = scoped_environments(options)
-                  accounts.each do |account|
-                    acc_map[account.links.self.href] = account
-                  end
+                  acc_map = environment_map(accounts)
 
                   Aptible::Api::MetricDrain.all(
                     token: fetch_token,