--- a/spec/aptible/cli/subcommands/log_drain_spec.rb
+++ b/spec/aptible/cli/subcommands/log_drain_spec.rb
@@ -10,8 +10,13 @@ describe Aptible::CLI::Agent do
   before { allow(subject).to receive(:fetch_token).and_return(token) }
 
   before do
+    allow(Aptible::Api::LogDrain).to receive(:all)
+      .with(token: token, href: '/log_drains?per_page=5000')
+      .and_return([log_drain])
+
     allow(Aptible::Api::Account).to receive(:all)
-      .with(token: token).and_return([account])
+      .with(token: token, href: '/accounts?per_page=5000&no_embed=true')
+      .and_return([account])
   end
 
   describe '#log_drain:list' do
@@ -25,8 +30,16 @@ describe Aptible::CLI::Agent do
 
     it 'lists log drains across multiple accounts' do
       other_account = Fabricate(:account)
-      Fabricate(:log_drain, handle: 'test2', account: other_account)
+      other_drain = Fabricate(
+        :log_drain,
+        handle: 'test2',
+        account: other_account
+      )
       accounts = [account, other_account]
+
+      allow(Aptible::Api::LogDrain).to receive(:all)
+        .with(token: token, href: '/log_drains?per_page=5000')
+        .and_return([log_drain, other_drain])
       allow(Aptible::Api::Account).to receive(:all).and_return(accounts)
 
       subject.send('log_drain:list')