--- a/spec/aptible/cli/subcommands/services_spec.rb
+++ b/spec/aptible/cli/subcommands/services_spec.rb
@@ -6,7 +6,9 @@ describe Aptible::CLI::Agent do
 
   before do
     allow(subject).to receive(:fetch_token) { token }
-    allow(Aptible::Api::App).to receive(:all).with(token: token)
+    allow(Aptible::Api::App)
+      .to receive(:all)
+      .with(token: token, href: '/apps?per_page=5000&no_embed=true')
       .and_return([app])
   end