https://app.shortcut.com/aptible/story/30651/cli-performance-for-list-commands

# environment:list

## before

```bash
$ time bundle exec bin/aptible environment:list

________________________________________________________
Executed in    2.84 secs      fish           external
   usr time  426.77 millis    0.25 millis  426.53 millis
   sys time  139.25 millis    1.70 millis  137.55 millis

```

## after

```bash
$ time bundle exec bin/aptible environment:list

________________________________________________________
Executed in    2.93 secs      fish           external
   usr time  383.18 millis    0.24 millis  382.94 millis
   sys time  142.15 millis    1.64 millis  140.51 millis
```

# apps

## before

```bash
$ time bundle exec bin/aptible apps

________________________________________________________
Executed in   14.55 secs    fish           external
   usr time    1.26 secs    0.26 millis    1.26 secs
   sys time    0.21 secs    1.70 millis    0.21 secs
```

## after

```bash
$ time bundle exec bin/aptible apps

________________________________________________________
Executed in    5.86 secs      fish           external
   usr time  600.29 millis    0.23 millis  600.06 millis
   sys time  200.44 millis    1.54 millis  198.90 millis
```

# db:list

## before

```bash
$ time bundle exec bin/aptible db:list

________________________________________________________
Executed in   47.10 secs    fish           external
   usr time    2.00 secs    0.24 millis    2.00 secs
   sys time    0.38 secs    1.75 millis    0.38 secs
```

## after

```bash
$ time bundle exec bin/aptible db:list

________________________________________________________
Executed in    4.86 secs      fish           external
   usr time  495.17 millis    0.22 millis  494.95 millis
   sys time  143.12 millis    1.48 millis  141.64 millis
```