@neurosnap 2025-02-07
Nice catch!  I missed a couple of account checks, should be fixed now.