@neurosnap 2025-02-10
Added an environment variable `APTIBLE_DEBUG` to set the log level for `aptible-resource`